Warning fixes in yuni-docmake
git-svn-id: file:///home/michael/svn/yuni/trunk@1747 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/tools/yuni-docmake/convert.cpp
@@ -299,7 +299,7 @@ namespace // anonymous
 			Compound::Ptr compound = pStack.front();
 
 			const TIXML_STRING& name = text.ValueTStr();
-			compound->name.append(name.c_str(), name.size());
+			compound->name.append(name.c_str(), (unsigned int)name.size());
 		}
 		return true;
 	}
